+++ src/Components/Components/src/Lifetime/ComponentApplicationState.cs
@@ -1,11 +1,17 @@
+// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
 using System;
 using System.Collections.Generic;
-using System.Collections.ObjectModel;
 using System.Diagnostics.CodeAnalysis;
+using System.Text.Json;
 using System.Threading.Tasks;
 
 namespace Microsoft.AspNetCore.Components
 {
+    /// <summary>
+    /// The state for the component and services of a components application.
+    /// </summary>
     public class ComponentApplicationState
     {
         private IDictionary<string, byte[]>? _existingState;
@@ -29,11 +35,21 @@ namespace Microsoft.AspNetCore.Components
             _existingState = existingState ?? throw new ArgumentNullException(nameof(existingState));
         }
 
+        /// <summary>
+        /// Registers a callback that will be invoked when the component application is being paused.
+        /// </summary>
+        /// <param name="callback">The <see cref="Func{TResult}"/> to invoke.</param>
         public void RegisterOnPersistingCallback(Func<Task> callback)
         {
             _registeredCallbacks.Add(callback);
         }
 
+        /// <summary>
+        /// Tries to retrieve the persisted state with the given <paramref name="key"/>.
+        /// </summary>
+        /// <param name="key">The key used to persist the state.</param>
+        /// <param name="value">The persisted state.</param>
+        /// <returns><c>true</c> if the state was found;<c>false</c> otherwise.</returns>
         public bool TryRetrievePersistedState(string key, [MaybeNullWhen(false)] out byte[] value)
         {
             if (_existingState == null)
@@ -44,77 +60,46 @@ namespace Microsoft.AspNetCore.Components
             return _existingState.TryGetValue(key, out value);
         }
 
+        /// <summary>
+        /// Persists the serialized state <paramref name="value"/> for the given <paramref name="key"/>.
+        /// </summary>
+        /// <param name="key">The key to use to persist the state.</param>
+        /// <param name="value">The state to persist.</param>
         public void PersistState(string key, byte[] value)
         {
             _currentState[key] = value;
         }
-    }
 
-    public class ComponentApplicationLifetime
-    {
-        private bool _stateIsPersisted;
-        private bool _pauseInProgress;
-        private List<Func<Task>> _pauseCallbacks = new();
-        private readonly Dictionary<string, byte[]> _currentState = new();
-
-        public ComponentApplicationLifetime()
+        /// <summary>
+        /// Serializes <paramref name="instance"/> as JSON and persists it under the given <paramref name="key"/>.
+        /// </summary>
+        /// <typeparam name="T">The <paramref name="instance"/> type.</typeparam>
+        /// <param name="key">The key to use to persist the state.</param>
+        /// <param name="instance">The instance to persist.</param>
+        public void PersistAsJson<T>(string key, T instance)
         {
-            State = new ComponentApplicationState(_currentState, _pauseCallbacks);
+            PersistState(key, JsonSerializer.SerializeToUtf8Bytes(instance));
         }
 
-        public ComponentApplicationState State { get; }
-
-        public Task RestoreStateAsync(IComponentApplicationStateStore store)
+        /// <summary>
+        /// Tries to retrieve the persisted state as JSON with the given <paramref name="key"/> and deserializes it into an
+        /// instance of type <typeparamref name="T"/>.
+        /// </summary>
+        /// <param name="key">The key used to persist the instance.</param>
+        /// <param name="instance">The persisted instance.</param>
+        /// <returns><c>true</c> if the state was found;<c>false</c> otherwise.</returns>
+        public bool TryRetrieveFromJson<T>(string key, [MaybeNullWhen(false)] out T instance)
         {
-            var data = store.GetPersistedState();
-            State.InitializeExistingState(data);
-
-            return Task.CompletedTask;
+            if (TryRetrievePersistedState(key, out var data))
+            {
+                instance = JsonSerializer.Deserialize<T>(data)!;
+                return true;
+            }
+            else
+            {
+                instance = default(T);
+                return false;
+            }
         }
-
-        public async Task PauseAsync()
-        {
-            _pauseInProgress = true;
-            foreach (var callback in _pauseCallbacks)
-            {
-                await callback();
-            }
-            _pauseInProgress = false;
-        }
-
-        public Task PersistStateAsync(IComponentApplicationStateStore store)
-        {
-            if (State == null)
-            {
-                throw new InvalidOperationException("ComponentApplicationLifetimeNotInitialized.");
-            }
-            if (_stateIsPersisted)
-            {
-                throw new InvalidOperationException("State already persisted.");
-            }
-            if (_pauseInProgress)
-            {
-                throw new InvalidOperationException("A pause operation is in progress.");
-            }
-            if (_pauseCallbacks.Count > 0)
-            {
-                throw new InvalidOperationException("Not all registered instances have been paused.");
-            }
-            _stateIsPersisted = true;
-            var data = new ReadOnlyDictionary<string, byte[]>(_currentState);
-            return store.PersistStateAsync(data);
-        }
-
-        internal void AddRegistration(Func<Task> callback)
-        {
-            _pauseCallbacks.Add(callback);
-        }
-    }
-
-    public interface IComponentApplicationStateStore
-    {
-        IDictionary<string, byte[]> GetPersistedState();
-
-        Task PersistStateAsync(IReadOnlyDictionary<string, byte[]> manager);
     }
 }
